+++ src/test/csrc/common/ram.h
@@ -31,8 +31,8 @@ long get_ram_size();
 void* get_img_start();
 long get_img_size();
 
-uint64_t pmem_read(uint64_t raddr);
-void pmem_write(uint64_t waddr, uint64_t wdata);
+uint32_t pmem_read(uint32_t raddr);
+void pmem_write(uint32_t waddr, uint32_t wdata);
 
 #ifdef WITH_DRAMSIM3
 #include "axi4.h"
